refactor checkbox
