+++ src/components/common/FormBuilder/QuestionBuilder/Checkbox/index.js
@@ -1,0 +1,4 @@
+export { default as Radio } from './Radio';
+export { default as RadioElse } from './RadioElse';
+export { default as Checkbox } from './Checkbox';
+export { default as CheckboxElse } from './CheckboxElse';
